+++ main.py
@@ -1,16 +1,12 @@
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from routes import activities, itinerary, default, saving
+from routes import activities, itinerary, default, saving, map
+from config import PORT
 from dotenv import load_dotenv
-import os
-from map import get_google_directions, DirectionsRequest
+
 
 # Load environment variables
 load_dotenv(override=True)
-
-# Configuration
-# BACKEND_URL = os.getenv("BACKEND_URL")
-PORT = int(os.getenv("PORT", "5000"))
 
 # Create FastAPI app
 app = FastAPI(title="Travelator Database API")
@@ -34,21 +30,8 @@ app.add_middleware(
 # Include routers
 app.include_router(activities.router)
 app.include_router(itinerary.router)
-
-
-@app.post("/get-directions")
-async def get_directions(request: DirectionsRequest):
-    route_data = get_google_directions(
-        request.origin, request.destination, request.mode
-    )
-    if not route_data:
-        return {"error": "No route found"}
-
-    return {"routes": route_data}
-
-
 app.include_router(saving.router)
-
+app.include_router(map.router)
 app.include_router(default.router)
 
 if __name__ == "__main__":
