+++ requirements.txt
@@ -62,3 +62,24 @@ virtualenv==20.29.2
 watchfiles==1.0.4
 websockets>=11,<15
 xyzservices==2025.1.0
+aiohappyeyeballs==2.4.8
+aiohttp==3.11.13
+aiosignal==1.3.2
+async-timeout==5.0.1
+attrs==25.1.0
+deprecation==2.1.0
+frozenlist==1.5.0
+gotrue==2.11.4
+h2==4.2.0
+hpack==4.1.0
+hyperframe==6.1.0
+multidict==6.1.0
+postgrest==0.19.3
+propcache==0.3.0
+python-dateutil==2.9.0.post0
+realtime==2.4.1
+six==1.17.0
+storage3==0.11.3
+StrEnum==0.4.15
+supafunc==0.9.3
+yarl==1.18.3
